Revert "Nightly -> Main release"
