+++ worker/agents/core/state.ts
@@ -1,11 +1,9 @@
-import type { PhasicBlueprint, AgenticBlueprint, PhaseConceptType ,
+import type { Blueprint, PhaseConceptType ,
     FileOutputType,
-    Blueprint,
 } from '../schemas';
 // import type { ScreenshotData } from './types';
 import type { ConversationMessage } from '../inferutils/common';
 import type { InferenceContext } from '../inferutils/config.types';
-import { BehaviorType, Plan } from './types';
 
 export interface FileState extends FileOutputType {
     lastDiff: string;
@@ -26,92 +24,33 @@ export enum CurrentDevState {
 
 export const MAX_PHASES = 12;
 
-/** Common state fields for all agent behaviors */
-export interface BaseProjectState {
-    behaviorType: BehaviorType;
-    // Identity
-    projectName: string;
-    query: string;
-    sessionId: string;
-    hostname: string;
-
+export interface CodeGenState {
     blueprint: Blueprint;
-
-    templateName: string | 'custom';
-    
-    // Conversation
-    conversationMessages: ConversationMessage[];
-    
-    // Inference context
-    inferenceContext: InferenceContext;
-    
-    // Generation control
-    shouldBeGenerating: boolean;
-    // agentMode: 'deterministic' | 'smart';    // Would be migrated and mapped to behaviorType
-    
-    // Common file storage
+    projectName: string,
+    query: string;
     generatedFilesMap: Record<string, FileState>;
-    
-    // Common infrastructure
+    generatedPhases: PhaseState[];
+    commandsHistory?: string[]; // History of commands run
+    lastPackageJson?: string; // Last package.json file contents
+    templateName: string;
     sandboxInstanceId?: string;
-    commandsHistory?: string[];
-    lastPackageJson?: string;
-    pendingUserInputs: string[];
-    projectUpdatesAccumulator: string[];
     
-    // Deep debug
-    lastDeepDebugTranscript: string | null;
-
+    shouldBeGenerating: boolean; // Persistent flag indicating generation should be active
     mvpGenerated: boolean;
     reviewingInitiated: boolean;
-}
-
-/** Phasic agent state */
-export interface PhasicState extends BaseProjectState {
-    behaviorType: 'phasic';
-    blueprint: PhasicBlueprint;
-    generatedPhases: PhaseState[];
-    
+    agentMode: 'deterministic' | 'smart';
+    sessionId: string;
+    hostname: string;
     phasesCounter: number;
+
+    pendingUserInputs: string[];
     currentDevState: CurrentDevState;
-    reviewCycles?: number;
-    currentPhase?: PhaseConceptType;
-}
+    reviewCycles?: number; // Number of review cycles for code review phase
+    currentPhase?: PhaseConceptType; // Current phase being worked on
+    
+    conversationMessages: ConversationMessage[];
+    projectUpdatesAccumulator: string[];
+    inferenceContext: InferenceContext;
 
-export interface WorkflowMetadata {
-    name: string;
-    description: string;
-    params: Record<string, {
-        type: 'string' | 'number' | 'boolean' | 'object';
-        description: string;
-        example?: unknown;
-        required: boolean;
-    }>;
-    bindings?: {
-        envVars?: Record<string, {
-            type: 'string';
-            description: string;
-            default?: string;
-            required?: boolean;
-        }>;
-        secrets?: Record<string, {
-            type: 'secret';
-            description: string;
-            required?: boolean;
-        }>;
-        resources?: Record<string, {
-            type: 'kv' | 'r2' | 'd1' | 'queue' | 'ai';
-            description: string;
-            required?: boolean;
-        }>;
-    };
-}
-
-/** Agentic agent state */
-export interface AgenticState extends BaseProjectState {
-    behaviorType: 'agentic';
-    blueprint: AgenticBlueprint;
-    currentPlan: Plan;
-}
-
-export type AgentState = PhasicState | AgenticState;
+    lastDeepDebugTranscript: string | null;
+}
