+++ worker/agents/services/implementations/DeploymentManager.ts
@@ -262,11 +262,12 @@ export class DeploymentManager extends BaseAgentService implements IDeploymentMa
      */
     async fetchRuntimeErrors(clear: boolean = true): Promise<RuntimeError[]> {
         const { sandboxInstanceId } = this.getState();
+        const logger = this.getLog();
+        const client = this.getClient();
+
         if (!sandboxInstanceId) {
             throw new Error('No sandbox instance available for runtime error fetching');
         }
-        const logger = this.getLog();
-        const client = this.getClient();
 
         const resp = await client.getInstanceErrors(sandboxInstanceId, clear);
