+++ worker/agents/services/implementations/FileManager.ts
@@ -133,9 +133,8 @@ export class FileManager implements IFileManager {
             if (shouldCommit) {
                 // If commit message is available, commit, else stage
                 if (commitMessage) {
-                    const unescapedMessage = commitMessage.replace(/\\n/g, '\n').replace(/\\t/g, '\t');
-                    console.log(`[FileManager] Committing ${fileStates.length} files:`, unescapedMessage);
-                    await this.git.commit(fileStates, unescapedMessage);
+                    console.log(`[FileManager] Committing ${fileStates.length} files:`, commitMessage);
+                    await this.git.commit(fileStates, commitMessage);
                     console.log(`[FileManager] Commit successful`);
                 } else {
                     console.log(`[FileManager] Staging ${fileStates.length} files`);
