+++ worker/agents/tools/toolkit/alter-blueprint.ts
@@ -1,6 +1,6 @@
 import { ToolDefinition } from '../types';
 import { StructuredLogger } from '../../../logger';
-import { ICodingAgent } from 'worker/agents/services/interfaces/ICodingAgent';
+import { CodingAgentInterface } from 'worker/agents/services/implementations/CodingAgent';
 import { Blueprint } from 'worker/agents/schemas';
 
 type AlterBlueprintArgs = {
@@ -10,7 +10,7 @@ type AlterBlueprintArgs = {
 };
 
 export function createAlterBlueprintTool(
-  agent: ICodingAgent,
+  agent: CodingAgentInterface,
   logger: StructuredLogger
 ): ToolDefinition<AlterBlueprintArgs, Blueprint> {
   return {
