+++ worker/agents/tools/toolkit/regenerate-file.ts
@@ -1,6 +1,6 @@
 import { ToolDefinition, ErrorResult } from '../types';
 import { StructuredLogger } from '../../../logger';
-import { ICodingAgent } from 'worker/agents/services/interfaces/ICodingAgent';
+import { CodingAgentInterface } from 'worker/agents/services/implementations/CodingAgent';
 
 export type RegenerateFileArgs = {
 	path: string;
@@ -12,7 +12,7 @@ export type RegenerateFileResult =
 	| ErrorResult;
 
 export function createRegenerateFileTool(
-	agent: ICodingAgent,
+	agent: CodingAgentInterface,
 	logger: StructuredLogger,
 ): ToolDefinition<RegenerateFileArgs, RegenerateFileResult> {
 	return {
