+++ worker/services/rate-limit/config.ts
@@ -81,9 +81,9 @@ export const DEFAULT_RATE_LIMIT_SETTINGS: RateLimitSettings = {
 	llmCalls: {
 		enabled: true,
 		store: RateLimitStore.DURABLE_OBJECT,
-		limit: 500,
+		limit: 800,
 		period: 60 * 60, // 1 hour
-        dailyLimit: 1700,
+        dailyLimit: 2000,
 		excludeBYOKUsers: true,
 	},
 };
